# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: DataCloak Sentiment Workbench API
  description: |
    Comprehensive API for sentiment analysis, data security, PII detection, and compliance management.
    
    ## Features
    - **Sentiment Analysis**: Advanced sentiment analysis with OpenAI integration
    - **Data Security**: PII detection, masking, and security auditing  
    - **Compliance**: GDPR, HIPAA, and custom compliance framework support
    - **Real-time Processing**: WebSocket and Server-Sent Events for live updates
    - **Batch Processing**: High-volume data processing with job queues
    - **Stream Processing**: Large file handling with memory-efficient streaming
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    Admin endpoints support both JWT and Basic authentication.
  version: 1.0.0
  contact:
    name: DataCloak API Support
    email: support@datacloak.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.datacloak-sentiment.com
    description: Production server
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      expiresIn:
                        type: string
                        example: 24h
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/verify:
    post:
      tags:
        - Authentication
      summary: Verify JWT token
      description: Verify the validity of a JWT token
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      valid:
                        type: boolean
                        example: true
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Sentiment Analysis
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication for admin endpoints
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
          example: admin
        role:
          type: string
          enum: [admin, user]
          example: admin
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Invalid request parameters
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Authentication required
            code: AUTH_REQUIRED
            timestamp: "2024-01-15T10:30:00Z"
...